M02A01 – Operações Aritméticas

+++ Modulo_02_Operadores/M02A01–Operacoes_Aritmeticas.c
@@ -5,7 +5,7 @@ void main(){
     setlocale(LC_ALL,"Portuguese");
 
     int a = 3;
-    int b = 2;
+    float b = 2;
     float r = a / b;
 
     printf("O resultado e %.2f", r);
